requirements.txt
uvicorn
fastapi
opencv-python-headless
numpy
pandas
python-multipart
tensorflow-cpu
streamlit_drawable_canvas
scikit-learn